feat(hub-common): move isLegacySearchCategory and toCollectionKey into internal helper folders
affects: @esri/hub-common
