+++ packages/common/src/search/_internal/commonHelpers/isLegacySearchCategory.ts
@@ -1,0 +1,17 @@
+export type LegacySearchCategory =
+  | "Site"
+  | "Event"
+  | "Dataset"
+  | "Document"
+  | "App,Map";
+
+export function isLegacySearchCategory(value: any) {
+  const categories: LegacySearchCategory[] = [
+    "Site",
+    "Event",
+    "Dataset",
+    "Document",
+    "App,Map",
+  ];
+  return categories.includes(value);
+}
