+++ packages/common/src/search/_internal/commonHelpers/toCollectionKey.ts
@@ -1,0 +1,7 @@
+import { LegacySearchCategory } from "./isLegacySearchCategory";
+
+export function toCollectionKey(legacySearchCategory: LegacySearchCategory) {
+  return legacySearchCategory === "App,Map"
+    ? "appAndMap"
+    : legacySearchCategory.toLowerCase();
+}
